Removes the last references to planning from runtime
Some tests were moved or split into a new integration-test directory src/tests.
