+++ src/planning/test/plan/planning-result-test.ts
@@ -11,13 +11,13 @@ import {assert} from '../../../platform/chai-web.js';
 import {Recipe} from '../../../runtime/recipe/recipe.js';
 import {Search} from '../../../runtime/recipe/search.js';
 import {Relevance} from '../../../runtime/relevance.js';
-import {TestHelper} from '../../../runtime/testing/test-helper.js';
+import {PlanningTestHelper} from '../../testing/planning-test-helper.js';
 import {PlanningResult} from '../../plan/planning-result.js';
 import {Suggestion} from '../../plan/suggestion.js';
 
 describe('planning result', () => {
   async function testResultSerialization(manifestFilename) {
-    const helper = await TestHelper.createAndPlan({manifestFilename});
+    const helper = await PlanningTestHelper.createAndPlan({manifestFilename});
     assert.isNotEmpty(helper.suggestions);
     helper.suggestions.forEach(s => {
       s.relevance = Relevance.create(helper.arc, s.plan);
@@ -38,7 +38,7 @@ describe('planning result', () => {
   });
 
   it('appends search suggestions', async () => {
-    const helper = await TestHelper.createAndPlan(
+    const helper = await PlanningTestHelper.createAndPlan(
         {manifestFilename: './src/runtime/test/artifacts/Products/Products.recipes'});
     const result = new PlanningResult(helper.envOptions);
     // Appends new suggestion.
@@ -100,7 +100,7 @@ recipe R3
     thing <- thingHandle
         `;
   async function prepareMerge(manifestStr1, manifestStr2) {
-    const helper = await TestHelper.create();
+    const helper = await PlanningTestHelper.create();
 
     const planToSuggestion = async (plan: Recipe): Promise<Suggestion> => {
       const suggestion = Suggestion.create(plan, await plan.digest(), Relevance.create(helper.arc, plan));
@@ -113,7 +113,7 @@ recipe R3
       return suggestion;
     };
     const manifestToResult = async (manifestStr) =>  {
-      const manifest = await TestHelper.parseManifest(manifestStr, helper.loader);
+      const manifest = await PlanningTestHelper.parseManifest(manifestStr, helper.loader);
       const result = new PlanningResult(helper.envOptions);
 
       const suggestions: Suggestion[] = await Promise.all(
