+++ src/runtime/test/particle-api-test.ts
@@ -833,8 +833,8 @@ describe('particle-api', () => {
     const [innerArc] = arc.findInnerArcs(transformationParticle);
 
     assert.equal(innerArc.activeRecipe.toString(), `recipe
-  slot '!493604905418752:demo:inner2:1' as slot0
-  slot 'slotid-!493604905418752:demo:inner2:2' as slot1
+  slot '!878110499405824:demo:inner2:1' as slot0
+  slot 'slotid-!878110499405824:demo:inner2:2' as slot1
   A as particle0
     consume content as slot0
       provide detail as slot1
