+++ src/tests/particles/artifacts/products-test.recipes
@@ -6,12 +6,12 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import '../../artifacts/Common/List.manifest'
-import '../../artifacts/Products/test/ProductFilter.manifest'
-import '../../artifacts/Products/test/ProductIsBook.manifest'
+import '../../../runtime/test/artifacts/Common/List.manifest'
+import '../../../runtime/test/artifacts/Products/test/ProductFilter.manifest'
+import '../../../runtime/test/artifacts/Products/test/ProductIsBook.manifest'
 
-import '../../artifacts/Products/Product.schema'
-particle ShowProduct in '../../artifacts/Products/source/ShowProduct.js'
+import '../../../runtime/test/artifacts/Products/Product.schema'
+particle ShowProduct in '../../../runtime/test/artifacts/Products/source/ShowProduct.js'
   in Product product
   consume item
