+++ src/tests/particles/artifacts/source/ShowFoo.js
