+++ src/tests/slot-composer-tests.ts
@@ -19,6 +19,7 @@ import {HostedSlotContext, ProvidedSlotContext} from '../runtime/slot-context.js
 import {MockSlotComposer} from '../runtime/testing/mock-slot-composer.js';
 import {StubLoader} from '../runtime/testing/stub-loader.js';
 import {TestHelper} from '../runtime/testing/test-helper.js';
+import {PlanningTestHelper} from '../planning/testing/planning-test-helper.js';
 
 async function initSlotComposer(recipeStr) {
   const slotComposer = new MockSlotComposer().newExpectations();
@@ -121,8 +122,8 @@ recipe
       .expectRenderSlot('ShowProduct', 'item', {'contentTypes': ['template', 'model', 'templateName']})
       .expectRenderSlot('ItemMultiplexer', 'item', {'contentTypes': ['template', 'model', 'templateName']});
 
-    const helper = await TestHelper.createAndPlan({
-      manifestFilename: './src/runtime/test/particles/artifacts/products-test.recipes',
+    const helper = await PlanningTestHelper.createAndPlan({
+      manifestFilename: './src/tests/particles/artifacts/products-test.recipes',
       slotComposer
     });
 
@@ -217,7 +218,7 @@ recipe
       .expectRenderSlot('TransformationParticle', 'root', {'contentTypes': ['model', 'templateName']})
       .expectRenderSlot('B', 'detail', {'contentTypes': ['model', 'templateName']});
 
-    const {arc} = await TestHelper.create({
+    const {arc} = await PlanningTestHelper.create({
       manifestString: `
         particle TransformationParticle in 'TransformationParticle.js'
           consume root
